requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.13
djangorestframework==3.14.0
drf-yasg==1.21.5
flake8==6.0.0
httpie==3.2.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pip==23.0.1
platformdirs==3.1.0
pluggy==1.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
PySocks==1.7.1
pytest==7.2.2
pytz==2022.7.1
requests-toolbelt==0.10.1
requests==2.28.2
rich==13.3.1
ruamel.yaml==0.17.21
setuptools==65.5.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14